Use shared LayoutDelegate for position, and move Content to desktop TST file
